+++ Revo.Core/Transactions/IUnitOfWorkAccessor.cs
@@ -1,0 +1,7 @@
+﻿namespace Revo.Core.Transactions
+{
+    public interface IUnitOfWorkAccessor
+    {
+        IUnitOfWork UnitOfWork { get; }
+    }
+}