+++ Revo.Infrastructure/Repositories/IRepositoryFactory.cs
@@ -1,10 +1,9 @@
-﻿using Revo.Core.Events;
-using Revo.Core.Transactions;
+﻿using Revo.Core.Transactions;
 
 namespace Revo.Infrastructure.Repositories
 {
     public interface IRepositoryFactory
     {
-        IRepository CreateRepository(IUnitOfWork unitOfWork);
+        IRepository CreateRepository(IUnitOfWorkAccessor unitOfWorkAccessor);
     }
 }